// dot-net-demo/dot-net-demo/src/src/app/models/license.ts
export class License {
  idLicense: number;
  status: string;
  maintainer: string;
  family: string;
  title: string;
  domain_data: boolean;
  is_okd_compliant: boolean;
  is_generic: boolean;
  url: string;
  is_osi_compliant: boolean;
  domain_content: boolean;
  domain_software: boolean;
  id: string;
}